// WADNR.Web/src/app/shared/generated/model/treatment-update-detail.ts
export class TreatmentUpdateDetail { 
    TreatmentUpdateID?: number;
    ProjectUpdateBatchID?: number;
    ProjectLocationUpdateID?: number | null;
    TreatmentAreaName?: string | null;
    TreatmentTypeID?: number;
    TreatmentTypeName?: string | null;
    TreatmentDetailedActivityTypeID?: number;
    TreatmentDetailedActivityTypeName?: string | null;
    TreatmentCodeID?: number | null;
    TreatmentCodeName?: string | null;
    TreatmentStartDate?: string | null;
    TreatmentEndDate?: string | null;
    TreatmentFootprintAcres?: number;
    TreatmentTreatedAcres?: number | null;
    CostPerAcre?: number | null;
    TotalCost?: number | null;
    TreatmentNotes?: string | null;
    ProgramID?: number | null;
    ProgramName?: string | null;
    ImportedFromGis?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}