// WADNR.Web/src/app/shared/generated/model/treatment-area-update-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentAreaUpdateLookupItem { 
    ProjectLocationUpdateID?: number;
    ProjectLocationUpdateName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentAreaUpdateLookupItemForm { 
    ProjectLocationUpdateID?: FormControl<number>;
    ProjectLocationUpdateName?: FormControl<string>;
}

export class TreatmentAreaUpdateLookupItemFormControls { 
    public static ProjectLocationUpdateID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationUpdateName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}